Fix security error with new Cordova app.
The content security policy that is now added to index.html for Cordova apps prevents websocket connections. As a temporary fix, add "ws:" to default-src. The way we're doing this is a temporary hack and should be replaced with a more robust solution (for example, this could be done during prepare for simulations: properly analyze and content security policy and add this item if it is not already there).